+++ sdk/keyvault/azure-security-keyvault-keys/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## 4.2.0-beta.3 (2020-04-09)
 - Added `LocalCryptographyClient`, `LocalCryptographyAsyncClient`, `LocalKeyEncryptionKeyClient` and `LocalKeyEncryptionKeyAsyncClient` to perform cryptography operations locally.
+- Added `retryPolicy` setter in `KeyClientBuilder`, `CryptographyClientBuilder` and `KeyEncryptionKeyClientBuilder`
 - Update azure-core dependency to version 1.4.0.
+
 
 ## 4.2.0-beta.2 (2020-03-10)
 ### Added